token_metadata: Speed up count_normal_token_owners
Currently, a set of nodes is built from _token_to_endpoint_map to get
the number of nodes in _token_to_endpoint_map.

To make it faster so we can call it on a fast path in the following
patch, a _nr_normal_token_owners member is introduced to track the
number.

Refs #10337
Refs #10817
Refs #10836
Refs #10837