locator: Speed up abstract_replication_strategy::get_address_ranges
To get the list of tokens for a given node, we loop through all the
tokens and calculate the nodes that are responsible for the token.

In case of the everywhere_topology, we know any node that is part of the
the ring will be responsible for all tokens.

This patch adds a fast path for everywhere_topology to avoid calculating
natural endpoints.

Refs #10337
Refs #10817
Refs #10836
Refs #10837